refactor: move dashboard lib
