+++ bin/cmds/spaces/reporting/delete.js
@@ -9,7 +9,7 @@ const get = require('lodash.get');
 
 const spacesLib = require('../../../../lib/spaces');
 const { findBySpace } = require('../../../../lib/reporting');
-const dashboard = require('../../../../lib/objects/dashboard');
+const dashboard = require('../../../../lib/dashboard');
 const reporting = require('../../../../lib/reporting');
 
 exports.command = 'delete [space]';
