+++ lib/dashboard.js
@@ -1,5 +1,5 @@
-const client = require('../app/elastic');
-const instance = require('../app/kibana');
+const client = require('./app/elastic');
+const instance = require('./app/kibana');
 
 const dashboard = {};
